Refactor: 리덕스 스토어 코드 분리
