+++ client/src/index.js
@@ -1,15 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux";
-import { configureStore } from "@reduxjs/toolkit";
-import { applyMiddleware, createStore } from "redux";
-import promiseMiddleware from "redux-promise";
-import ReduxThunk from "redux-thunk";
-import Reducer from "./_reducers";
+import reportWebVitals from "./reportWebVitals";
+import App from "./App";
+import store from "./store/store";
 
+import "./index.css";
 import { ThemeProvider, createTheme } from "@mui/material";
 
 import {
@@ -38,16 +34,6 @@ const muiTheme = createTheme({
 //   middleware: (getDefaultMiddleware) =>
 //     getDefaultMiddleware().concat(promiseMiddleware),
 // });
-
-const createStoreWithMiddleware = applyMiddleware(
-  promiseMiddleware,
-  ReduxThunk
-)(createStore);
-
-const store = createStoreWithMiddleware(
-  Reducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
